Shuffle startup code to pass through a separate command line parser. Add a NimbleApplication class that can be derived from so that alternate applications can use it. Rename nimble_main.h/.cc to nimble.h/.cc and make nimble_main actually contain the main function. Executables using NimbleSM as a library will have to inherit NimbleApplication and adjust command line parsing and subsystem created. Finally, move integrators to their own source files.
