+++ src/integrators/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(nimble PRIVATE
+        integrator_base.cc
+        explicit_time_integrator.cc
+        quasistatic_time_integrator.cc
+        )
