+++ src/nimble_cli.cc
@@ -43,16 +43,49 @@
 
 #include "nimble_cli.h"
 
-#include <CLI11/CLI11.hpp>
+#include "nimble_parser.h"
+
+#ifdef NIMBLE_HAVE_VT
+// From VT, if not we get some nasty conflicting symbols
+#include <CLI/CLI11.hpp>
+#else
+#include <nimble_tpl/CLI11/CLI11.hpp>
+#endif
 
 namespace nimble {
-int
-parseCommandLine(int argc, char** argv, nimble::EnvironmentFlags& myFlags)
-{
-  CLI::App nimble_app{"A Lagrangian finite-element code for solid mechanics."};
-  nimble_app.allow_extras();
 
-  auto* kokkos_opt = nimble_app.add_flag("--use_kokkos", myFlags.use_kokkos_, "Use Kokkos for parallel compute")
+struct CommandLineConfiguration::impl
+{
+  CLI::App cli_app_{"A Lagrangian finite-element code for solid mechanics."};
+  int argc_ = 0;
+  char **argv_ = nullptr;
+
+  bool use_kokkos_ = false;
+  bool use_tpetra_ = false;
+  bool use_vt_ = false;
+  bool use_uq_ = false;
+  std::string input_filename_;
+};
+
+CommandLineConfiguration::CommandLineConfiguration(int argc, char** argv)
+  : impl_{std::make_unique< impl >()}
+{
+  impl_->argc_ = argc;
+  impl_->argv_ = argv;
+}
+
+CommandLineConfiguration::CommandLineConfiguration(CommandLineConfiguration &&) noexcept = default;
+
+CommandLineConfiguration::~CommandLineConfiguration() = default;
+
+CommandLineConfiguration &
+CommandLineConfiguration::operator=(CommandLineConfiguration &&) noexcept = default;
+
+void
+CommandLineConfiguration::ConfigureCommandLineArguments() {
+  impl_->cli_app_.allow_extras();
+
+  auto* kokkos_opt = impl_->cli_app_.add_flag("--use_kokkos", impl_->use_kokkos_, "Use Kokkos for parallel compute")
                          ->check([](const std::string&) -> std::string {
 #ifdef NIMBLE_HAVE_KOKKOS
                            return {};  // Empty string on success
@@ -61,7 +94,7 @@ parseCommandLine(int argc, char** argv, nimble::EnvironmentFlags& myFlags)
 #endif
                          });
 
-  auto* tpetra_opt = nimble_app.add_flag("--use_tpetra", myFlags.use_tpetra_, "Use TPetra solvers")
+  auto* tpetra_opt = impl_->cli_app_.add_flag("--use_tpetra", impl_->use_tpetra_, "Use TPetra solvers")
                          ->check([](const std::string&) -> std::string {
 #ifdef NIMBLE_HAVE_TRILINOS
                            return {};  // Empty string on success
@@ -70,7 +103,7 @@ parseCommandLine(int argc, char** argv, nimble::EnvironmentFlags& myFlags)
 #endif
                          });
 
-  nimble_app.add_flag("--use_vt", myFlags.use_vt_, "Use DARMA/vt for asynchronous distributed contact tasking")
+  impl_->cli_app_.add_flag("--use_vt", impl_->use_vt_, "Use DARMA/vt for asynchronous distributed contact tasking")
           ->check([](const std::string&) -> std::string {
 #ifdef NIMBLE_HAVE_VT
             return {};  // Empty string on success
@@ -79,14 +112,71 @@ parseCommandLine(int argc, char** argv, nimble::EnvironmentFlags& myFlags)
 #endif
           });
 
+  impl_->cli_app_.add_flag("--use_uq", impl_->use_vt_, "Use UQ capability")->check([](const std::string&) -> std::string {
+#ifdef NIMBLE_HAVE_UQ
+    return {};  // Empty string on success
+#else
+    return "NimbleSM was not built with UQ support";
+#endif
+  });
+
+  impl_->cli_app_.add_option( "INPUT_FILE", impl_->input_filename_)->required()->check(CLI::ExistingFile);
+
   kokkos_opt->excludes(tpetra_opt);
   tpetra_opt->excludes(kokkos_opt);
+}
 
+int
+CommandLineConfiguration::ParseAndGetErrorCode() {
   try {
-    nimble_app.parse(argc, argv);
+    impl_->cli_app_.parse(impl_->argc_, impl_->argv_);
   } catch (const CLI::ParseError& e) {
-    return nimble_app.exit(e);
+    return impl_->cli_app_.exit(e);
   }
+
   return 0;
 }
+
+bool
+CommandLineConfiguration::UseKokkos() const noexcept
+{
+  return impl_->use_kokkos_;
+}
+
+bool
+CommandLineConfiguration::UseTpetra() const noexcept
+{
+  return impl_->use_tpetra_;
+}
+
+bool
+CommandLineConfiguration::UseVT() const noexcept
+{
+  return impl_->use_vt_;
+}
+
+bool
+CommandLineConfiguration::UseUQ() const noexcept
+{
+  return impl_->use_uq_;
+}
+
+const std::string &
+CommandLineConfiguration::InputFilename() const noexcept
+{
+  return impl_->input_filename_;
+}
+
+int &
+CommandLineConfiguration::ArgC() noexcept
+{
+  return impl_->argc_;
+}
+
+char **&
+CommandLineConfiguration::ArgV() noexcept
+{
+  return impl_->argv_;
+}
+
 }  // namespace nimble
